pyproject.toml
[tool.poetry]
name = "cirrus-deploy"
version = "0.1.0"
description = "Scripts for building and deploying OpenGoSim Cirrus on Equinor premises"
authors = ["Zohar Malamant <ZOM@equinor.com>"]
license = "GPLv3.0"
readme = "README.md"
packages = [{include="deploy", from="src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pydantic = "^2.9.2"
pyyaml = "^6.0.2"
six = "^1.16.0"
networkx = "^3.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
ruff = "^0.6.8"

[tool.poetry.scripts]
deploy = "deploy.__main__:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
